+++ src/constants.ts
@@ -1,2 +1,3 @@
 export const CONTEXT_UPDATE_DEBOUNCE_MS = 50;
 export const METADATA_FILENAME = '.oci-explorer.json';
+export const ORAS_COMMAND = 'oras';